[mod/#127] settingRoute 병합
